backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
openai>=1.30
python-dotenv>=1.0
python-multipart>=0.0.9
